+++ CHANGELOG.md
@@ -5,9 +5,9 @@ The format is inspired by [Keep aChangelog](https://keepachangelog.com/en/1.0.0/
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added Features and Improvements 🙌:
 - Add support for themed app icon (android 13)
+- Using latest flutter 3.7 version with improved Material You theme
 
 ## [0.2.2] - 2022-10-18
 ### Added Features and Improvements 🙌: